OrfFinder is a coding problem which involved using a suffix trie
The OrfFinder class takes in a specific string and the function "find" within the class takes in 2 arguments (prefix and a suffix) and returns all substrings which the OrfFinder class takes in with the specified prefix and suffix in a list. The problem was solved by using the Suffix Trie data structure.